+++ contracts/interface/ISettlementVault.sol
@@ -1,7 +1,4 @@
-
-
-
 interface ISettlementVault {
-    function getBalance(address user) external returns (uint, uint64);
+    function getBalance(address user) external returns (uint256, uint64);
     function accountToPaymentNonce(address account) external returns (uint16);
 }
